Crate onnxrt

onnxruntime_c_api.h

Macros

impl_AsONNXTensorElementDataType

Structs

Allocator

ArenaCfg

CudaProviderOptions

Env

IoBinding

MemoryInfo

ModelMetadata

OrtPrepackedWeightsContainer

PrepackedWeightsContainer

RunOptions

Session

SessionOptions

TensorTypeAndShapeInfo

ThreadingOptions

TypeInfo

UnownedMemoryInfo

UnownedTensorTypeAndShapeInfo

Value

Enums

Error

ExecutionMode

GraphOptimizationLevel

\brief Graph optimization level

ONNXTensorElementDataType

Copied from TensorProto::DataType Currently, Ort doesn’t support complex64, complex128

ONNXType

OrtAllocatorType

OrtErrorCode

OrtLanguageProjection

\brief Language projection identifiers /see OrtApi::SetLanguageProjection

OrtLoggingLevel

\brief Logging severity levels

OrtMemType

\brief Memory types for allocated memory, execution provider specific types should be extended in each provider.

OrtMemoryInfoDeviceType

\brief This mimics OrtDevice type constants so they can be returned in the API

Constants

ORT_API_VERSION

Statics

ALLOCATOR_WITH_DEFAULT_OPTIONS

ORT_API

Traits

AsONNXTensorElementDataType

Functions

available_providers

onnxruntime_c_api.h

build_info_string

onnxruntime_c_api.h

current_gpu_device_id

onnxruntime_c_api.h

execution_provider_api

onnxruntime_c_api.h

set_current_gpu_device_id

onnxruntime_c_api.h

version_string

onnxruntime_c_api.h

Type Definitions

Result