Crate onnxrt

onnxruntime_c_api.h

Macros

impl_AsONNXTensorElementDataType

Structs

Allocator

ArenaCfg

Env

IoBinding

MemoryInfo

ModelMetadata

RunOptions

Session

SessionOptions

TensorTypeAndShapeInfo

ThreadingOptions

TypeInfo

UnownedMemoryInfo

UnownedTensorTypeAndShapeInfo

Value

Enums

Error

ExecutionMode

GraphOptimizationLevel

ONNXTensorElementDataType

ONNXType

OrtAllocatorType

OrtErrorCode

OrtLanguageProjection

OrtLoggingLevel

OrtMemType
memory types for allocator, exec provider specific types should be extended in each provider Whenever this struct is updated, please also update the MakeKey function in onnxruntime/core/framework/execution_provider.cc

Constants

ORT_API_VERSION

Statics

ALLOCATOR_WITH_DEFAULT_OPTIONS

ORT_API

Traits

AsONNXTensorElementDataType

Functions

available_providers
onnxruntime_c_api.h

Type Definitions

Result