Struct onnxrt::Value

#[repr(transparent)]pub struct Value<'d> { /* fields omitted */ }

Implementations

impl<'d> Value<'d>

pub fn new_tensor(
shape: &[i64],
element_type: ONNXTensorElementDataType
) -> Result<Self>

pub fn new_tensor_using_allocator(
allocator: &mut Allocator,
shape: &[i64],
element_type: ONNXTensorElementDataType
) -> Result<Self>

onnxruntime_c_api.h

pub fn new_tensor_with_data<T: AsONNXTensorElementDataType>(
memory_info: &MemoryInfo,
data: &'d mut [T],
shape: &[i64]
) -> Result<Self>

onnxruntime_c_api.h

pub fn is_tensor(&self) -> bool

onnxruntime_c_api.h

pub fn tensor_data<T>(&self) -> Result<*const T>

onnxruntime_c_api.h

pub fn tensor_data_mut<T>(&mut self) -> Result<*mut T>

onnxruntime_c_api.h

pub unsafe fn get<T>(&self, indices: &[i64]) -> Result<&T>

onnxruntime_c_api.h

Safety

T must be the same element type as specified when self was created.

pub unsafe fn get_mut<T>(&mut self, indices: &[i64]) -> Result<&mut T>

onnxruntime_c_api.h

Safety

T must be the same element type as specified when self was created.

pub fn type_info(&self) -> Result<Option<TypeInfo>>

onnxruntime_c_api.h

pub fn tensor_type_info(&self) -> Result<TensorTypeAndShapeInfo>

onnxruntime_c_api.h

pub fn value_type(&self) -> Result<ONNXType>

onnxruntime_c_api.h